requirements.txt
alembic==1.10.2
certifi==2022.12.7
fastapi==0.95.0
fastapi-sqlalchemy==0.2.1
flower==1.2.0
nltk==3.8.1
pip-chill==1.0.1
psycopg2-binary==2.9.5
python-dotenv==1.0.0
redis==4.5.3
uvicorn==0.21.1
alembic==1.10.2
django-cors-headers==3.14.0
django-environ==0.10.0
djangorestframework==3.14.0
fastapi==0.95.0
fastapi-sqlalchemy==0.2.1
flower==1.2.0
fqdn==1.5.1
gensim==4.3.1
isoduration==20.11.0
jsonpointer==2.3
jupyter==1.0.0
openai==0.27.2
pip-chill==1.0.1
psycopg2-binary==2.9.5
python-dotenv==1.0.0
python-version==0.0.2
redis==4.5.3
scikit-learn==1.2.2
tensorflow==2.12.0
textblob==0.17.1
torchaudio==2.0.1
torchvision==0.15.1
transformers==4.27.3
tweepy==4.13.0
uri-template==1.2.0
uvicorn==0.21.1
vadersentiment==3.3.2
webcolors==1.12
